chore: 함수, 훅 위치 수정
