+++ src/api/domain/mypage/index.ts
@@ -1,0 +1,23 @@
+import { API_PATH } from "@api/constants/apiPath";
+import { get } from "@api/index";
+import { paths } from "@type/schema";
+
+//todo : try-catch문 작성하기
+export const getMemberInfo = async () => {
+  //type getMemberInfoRequest = paths["/api/dev/members"]["get"]["requestBody"];
+  type MemberInfoResponse = paths["/api/dev/members"]["get"]["responses"]["200"]["content"]["*/*"];
+  const { data } = await get<MemberInfoResponse>(API_PATH.MEMBERS, {});
+
+  return data.data;
+};
+
+//nickname을 안주고 accessToken만 주면 나의 반려동물 정보
+export const getPetInfo = async (nickname?: string) => {
+  type MyPetInfo = paths["/api/dev/pets"]["get"]["responses"]["200"]["content"]["*/*"];
+  const { data } = await get<MyPetInfo>(`${API_PATH.PETS}`, {
+    params: {
+      nickname,
+    },
+  });
+  return data.data;
+};
